Make SslBufferPool an interface
