+++ src/main/java/org/jboss/netty/handler/ssl/SslHandler.java
@@ -171,7 +171,7 @@ public class SslHandler extends FrameDecoder
      */
     public static synchronized SslBufferPool getDefaultBufferPool() {
         if (defaultBufferPool == null) {
-            defaultBufferPool = new SslBufferPool();
+            defaultBufferPool = new DefaultSslBufferPool();
         }
         return defaultBufferPool;
     }
